\name{nmScatterPlot}
\alias{nmScatterPlot}
\alias{nmScatterPlot}
\title{NONMEM scatter plot}
\description{Generates a set of scatter plots with features that are tailored for PK/PD data generated by NONMEM.}
\value{An object of class \code{multiTrellis} holding the plot}
\author{Mango Solutions}
\keyword{hplot}
\arguments{\item{obj}{An object of class NMRun, NMProblem, or data.frame. The object from which data will be plotted.}
\item{xVars}{Single string with name of x-axis variable.}
\item{yVars}{Variables to plot on the y-axes, specified as a comma seperated string or character vector}
\item{bVars}{“Trellis” variables on which to split data.}
\item{gVars}{“Grouping” variable – used to group points by colour, for legends and so on.}
\item{iVars}{Variable name of the “inherent grouping” (subject identifier) variable.  This affects plots whose type is "l", "i", "o" or "t".
Lines will be grouped by this variable in addition to any other grouping (e.g. in gVar), but colours will not vary in this grouping}
\item{addLegend}{Should legends be added?}
\item{addGrid}{should grids be added?}
\item{addLoess}{should a loess smoother line be added?}
\item{titles}{Plot title}
\item{logX}{Should the x-axis be logged?}
\item{logY}{similar to logX for y axis}
\item{idLines}{Should reference lines of slope 1 and y-intercept 0 be added?}
\item{abLines}{A list of reference lines.  Each element should be a length 1 or 2 numeric vector.  If length 1, a 
horizontal reference line will be added at that point, otherwise the intercept and slope of the line.  If parameter does not satisfy
these conditions, a warning will be emitted}
\item{xLab}{x-axis label.  By default, will use the names of the x-axis variable}
\item{yLab}{similar to the above, but for y-axis label}
\item{doPlot}{Should the plot be plotted immediately, before returning it as an object?}
\item{types}{Plot type to use.  Allowed types are "p" (for standard points), "l" (lines connected by the variable
specified in "iVars", "i" for points labeled by "iVars"), "o" (lines and points), and "t" for labels connected by lines grouped by "iVars"}
\item{overlaid}{Logical flag. If TRUE, for each fixed x, the y variables will be overlaid onto a single plot}
\item{problemNum}{Number of the problem (applicable to NMRun class only)}
\item{subProblems}{Number of the simulation subproblems to use (applicable to the NMSim* classes obly)}
\item{uniqueX}{Number of unique values for the X axis.  If number of uniques is less than or equal to this, a boxplot is created instead}
\item{yAxisScaleRelations}{Y-axis scale relations when panels are displayed. One of \code{"same"}, \code{"free"} or \code{"sliced"}.}
\item{maxPanels}{Maximum number of panels that should appear on each page of a graph.}
\item{xRotAngle}{Angle by which to rotate the x-axis tick marks}
\item{layout}{A length 2 vector which is passed in as the layout parameter to xyplot}
\item{maxTLevels}{If a single numeric (or string), the maximum number of levels that a "by" variable can have before it is binned.
If a character vector or a vector of length greater than one, the explicit breakpoints.}
\item{graphParams}{A full list of graphical parameter settings, such as that returned by getAllGraphParams().
Will override global settings. By default, these will be the global settings}
\item{xAxisScaleRelations}{X-axis scale relations when panels are displayed. One of \code{"same"}, \code{"free"} or \code{"sliced"}.}
\item{\dots}{Additonal variables passed to the xyplot function}
}
\examples{Theoph.df <- as.data.frame(Theoph)
nmScatterPlot(Theoph.df, xVar = "Time", yVar = "conc", iVar = "Subject", type = "l", title = "Theophiline", xLab = "conc") 
Indometh.df <- as.data.frame(Indometh)
nmScatterPlot(Indometh.df, xVar = "time", yVar = "conc", bVar = "Subject")       
}}

