\name{setAllGraphParams}
\alias{setAllGraphParams}
\title{Set all graphical parameters}
\usage{setAllGraphParams(settings)
}
\description{Set all graphical style options}
\value{None}
\author{fgochez}
\keyword{utilities}
\arguments{\item{settings}{A full list of graphical parameters.  See the documentation for RNMGraphics.}
}

