\name{Graphical parameters}
\alias{setGraphParams}
\alias{getGraphParams}
\alias{getAllGraphParams}

\title{Change and Retrieve Graphical Parameters}
\description{
  \code{setGraphParams} may be used to change graphical settings in a manner similar to the \code{lattice} package's
  function \code{trellis.par.set}, while \code{getGraphParams} is similar to \code{trellis.par.get}.
  Settings apply globally and to all plots generated by this package, and don't just apply to devices.
}
\usage{
setGraphParams(field, setting)
getGraphParams(field)
getAllGraphParams()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{field}{A string with the name of the field/option (e.g. \code{loess.line}) that the user wishes to change or retrieve.}
  \item{setting}{A named list which contains the new settings.}
}
\details{
	At the moment, the following fields exist:
  \itemize{
  \item \code{plot.symbol} - The plot points which appear in scatter plots.
  \item \code{superpose.symbol} - The plot points which appear in scatter plots whenever a grouping variable is used.
  \item \code{plot.line} - The plot lines which appear in scatter plots, whenever lines are chosen to connect data rather than points.
  \item \code{plot.text} - The plot labels that appear in scatter plots when type \code{"i"} is used.
  \item \code{loess.line} - The loess smoother curve which can be added to scatter plots.
  \item \code{histogram} - The fill colour and outline colour of histogram rectangles.
  \item \code{refline} - The reference lines that can be added to many plots.
  \item \code{barchart} - The fill colours of polygons used in \code{nmBarchart}.
  \item \code{axis.text} - The text that appears in y- and x-axis labels.
  \item \code{box.rectangle} - The rectangles in box plots.
  \item \code{title.text} - The plot main title text.
  \item \code{grid} - The reference/background grid text.
  \item \code{strip.bg} - Strip background settings
  \item \code{layout.heights} - Equivalent to the lattice setting of the same name
  \item \code{layout.widths} - Equivalent to the lattice setting of the same name
  \item \code{strip} - Strip function to use 
  \item \code{panelLayout} - Currently unused
  \item \code{panelMisc} - Miscellaneous controls over lattice panels 
  \item \code{legend} - Settings for the legend.
	}
}
\value{
  \code{setGraphParams} returns nothing.  \code{getGraphParams} returns a list with the configuration for the particular
  field/option chose, and \code{getAllGraphParams} returns a list (of lists) with all current configuration information.
}
\author{Mango Solutions <support@mango-solutions>}
\note{ Some of these settings and their implementations will almost certainly change in forthcoming releases.
}

\seealso{\code{\link{trellis.par.set}}}
\examples{
	nmScatterPlot(mtcars, xVars = "wt", yVars = "mpg, disp")
	plot.symbol <- getGraphParams("plot.symbol")
	old.plot.symbol <- plot.symbol
	plot.symbol$col <- "black"; plot.symbol$pch = 1
	setGraphParams("plot.symbol", plot.symbol)
	nmScatterPlot(mtcars, xVars = "wt", yVars = "mpg, disp")
	setGraphParams("plot.symbol", old.plot.symbol)
}

\keyword{ environment }
