% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmDotPlot.R
\name{nmDotPlot}
\alias{nmDotPlot}
\title{NONMEM dot plot}
\usage{
nmDotPlot(obj, factVar, contVar, bVars = NULL, iVar = "ID",
  gVar = "NULL", title = NULL, xLabs = NULL, yLabs = NULL,
  layout = NULL, maxPanels = numeric(0), addLegend = TRUE,
  maxTLevels = Inf, maxFactPerPanel = Inf, problemNum = 1,
  subProblems = 1, xAxisScaleRelations = c("same", "free", "sliced"),
  yAxisScaleRelations = c("same", "free", "sliced"),
  xAxisPlotStyle = c("data", "cont", "cat"), ...)
}
\arguments{
\item{obj}{An object of class \code{NMRun}, \code{NMProblem}, or \code{data.frame}. The object from which data will be plotted.}

\item{factVar}{Character variable representing the factor variable.}

\item{contVar}{Character vector or comma separated string specifying continuous variables.}

\item{bVars}{Trellis variables, specified as characters (or \code{NULL}, which is the default).}

\item{iVar}{Subject identifier variable (single string or NULL).}

\item{gVar}{Grouping variable}

\item{title}{Plot main title}

\item{xLabs}{X-axis label}

\item{yLabs}{Y-axis label}

\item{layout}{Numeric vector giving the number of columns, rows and pages in a multipanel display.  is overwritten by maxPanels.}

\item{maxPanels}{Maximum number of panels that should appear on each page of a graph.}

\item{addLegend}{Logical flag.  Should legends be added?}

\item{maxTLevels}{If a single numeric (or string), the maximum number of levels that a "by" variable can have before it is binned.
If a character vector or a vector of length greater than one, the explicit breakpoints.}

\item{maxFactPerPanel}{Max fact per panel.}

\item{problemNum}{The problem required for a \code{NMRun} object.}

\item{subProblems}{The sub problem of a run with simulations.}

\item{xAxisScaleRelations}{X-axis scale relations when panels are displayed. One of \code{"same"}, \code{"free"} or \code{"sliced"}.}

\item{yAxisScaleRelations}{Y-axis scale relations when panels are displayed. One of \code{"same"}, \code{"free"} or \code{"sliced"}.}

\item{xAxisPlotStyle}{"data" or "cont" or "cat".}

\item{...}{Additional parameters to \code{dotplot}.}
}
\value{
Multitrellis class object containing the plot.
}
\description{
Create a custom dotplot of one or more continuous variables against a categorical variable
}
\examples{
\dontrun{ 
 Theoph.df <- as.data.frame(Theoph)
 nmDotPlot( Theoph.df, factVar = "Subject", contVar = "Dose" )
}
test_data <- data.frame(X = 1:20, Y = 1:20, Z = c(1:10, seq(from = 11, to = 12, length.out = 10)),
W = 20:1, G = rep(LETTERS[1:2], 10), B = rep(letters[1:2], each = 10))
nmDotPlot(test_data, factVar = "B", contVar = "X,W")
}
\author{
Mango Solutions
}
\keyword{hplot}
